backend/requirements.txt
Django===5.2
djangorestframework===3.16.1
djangorestframework-simplejwt===5.5.1
psycopg2-binary===2.9.11
gunicorn===23.0.0
drf-spectacular===0.28.0
drf-spectacular-sidecar===2025.10.1
faker===37.11.0
